fix(marketplace): share sidebar autocomplete and soft-guard context
Extract EntityAutocompleteDropdown for glossary/marketplace search, and
no-op marketplace EntityProfile sync when the provider is absent.
